\name{katexTestPackage-package}
\alias{katexTestPackage-package}
\alias{katexTestPackage}
\docType{package}
\title{\packageTitle{katexTestPackage}}
\description{\packageDescription{katexTestPackage}}
\section{Package Content}{\packageIndices{katexTestPackage}}
\author{\packageAuthor{katexTestPackage}}
\section{Maintainer}{\packageMaintainer{katexTestPackage}}
\keyword{package}
