% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{A placeholder function}
\usage{
hello(txt = "world")
}
\arguments{
\item{txt}{An optional character variable, defaults to \sQuote{world}}
}
\value{
Nothing is returned but as a side effect output is printed
}
\description{
This function shows a standard text on the console. In a time-honoured
tradition, it defaults to displaying \emph{hello, world}.
}
\examples{
hello()
hello("and goodbye")
}
